+++ CHANGELOG.rst
@@ -4,8 +4,11 @@ CHANGELOG
 UNRELEASED
 ----------
 
-* 🎉 feat: categorization added
-* 🎉 feat: ignore folder feature added
+1.2.0 (2022-03-16)
+------------------
+
+* feat: categorization added
+* feat: ignore folder feature added
 
 1.1.0 (2022-03-14)
 ------------------